requirements.txt
django
paho-mqtt
arrow
numpy
psycopg2
ipdb
jupyter
django-extensions
